requirements.txt
colorama
python-dotenv
pyperclip
selenium